requirements.txt
fenics-dolfinx
mpich
alphashape
igraph
hdbscan
igraph
ipympl
ipykernel
jax
jupyter
libptscotch
libscotch
matplotlib
meshio
numpy
openpyxl
pandas
pybamm
python-gmsh
pyvista
pyvistaqt
scipy
scikit-learn
tetgen
vtk
